requirements.txt
streamlit>=1.30.0
pyyaml>=6.0
requests>=2.31.0
feedparser>=6.0
python-dateutil>=2.8
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=1.5.0
tiktoken>=0.5.0
plotly>=5.18.0
structlog>=23.1.0
diskcache>=5.6.0
httpx>=0.25.0
pytest>=7.4.0
pytest-cov>=4.1.0
python-dotenv>=1.0.0
tenacity>=8.2.0
xxhash>=3.4.0
